+++ packages/aws-lambda-graphql/src/__tests__/DynamoDBSubscriptionManager.test.ts
@@ -29,7 +29,7 @@ describe('DynamoDBSubscriptionManager', () => {
     transactWritePromiseMock.mockReset();
   });
 
-  describe('subscribersByEventName', () => {
+  describe('subscribersByEvent', () => {
     it('works correctly for emty query result', async () => {
       const subscriptionManager = new DynamoDBSubscriptionManager();
 
@@ -38,9 +38,10 @@ describe('DynamoDBSubscriptionManager', () => {
       let pages = 0;
 
       // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      for await (const page of subscriptionManager.subscribersByEventName(
-        'test',
-      )) {
+      for await (const page of subscriptionManager.subscribersByEvent({
+        event: 'test',
+        payload: {},
+      })) {
         pages++;
       }
 
@@ -68,9 +69,10 @@ describe('DynamoDBSubscriptionManager', () => {
       let pages = 0;
 
       // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      for await (const page of subscriptionManager.subscribersByEventName(
-        'test',
-      )) {
+      for await (const page of subscriptionManager.subscribersByEvent({
+        event: 'test',
+        payload: {},
+      })) {
         pages++;
       }