+++ packages/chat-example-server/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.0-alpha.19 (2020-11-13)
+
+**Note:** Version bump only for package serverless-chat-example-server
+
+## 1.0.0-alpha.18 (2020-09-18)
+
+**Note:** Version bump only for package serverless-chat-example-server
+
 ## 1.0.0-alpha.17 (2020-09-10)
 
 **Note:** Version bump only for package serverless-chat-example-server